+++ src/Type/Invoice.php
@@ -20,7 +20,7 @@ namespace MarcusJaschen\Collmex\Type;
  * @property $customer_title
  * @property $customer_firstname
  * @property $customer_lastname
- * @property $customer_firm
+ * @property $customer_company
  * @property $customer_department
  * @property $customer_street
  * @property $customer_zipcode
@@ -67,7 +67,7 @@ namespace MarcusJaschen\Collmex\Type;
  * @property $delivery_title
  * @property $delivery_firstname
  * @property $delivery_lastname
- * @property $delivery_firm
+ * @property $delivery_company
  * @property $delivery_department
  * @property $delivery_street
  * @property $delivery_zipcode
@@ -221,7 +221,7 @@ class Invoice extends AbstractType implements TypeInterface
         // 10
         'customer_firstname' => null,
         'customer_lastname' => null,
-        'customer_firm' => null,
+        'customer_company' => null,
         'customer_department' => null,
         'customer_street' => null,
         // 15
@@ -277,7 +277,7 @@ class Invoice extends AbstractType implements TypeInterface
         'delivery_title' => null,
         'delivery_firstname' => null,
         'delivery_lastname' => null,
-        'delivery_firm' => null,
+        'delivery_company' => null,
         // 60
         'delivery_department' => null,
         'delivery_street' => null,